requirements.txt
# Polymarket Notifier Bot Dependencies
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.3
aiohttp>=3.9.0

# Web scraping and browser automation
playwright>=1.40.0,<1.50.0

# Data processing
beautifulsoup4>=4.12.0

# Fix for Python 3.13 compatibility
greenlet>=3.0.0,<4.0.0

# Database (included with Python)
# sqlite3 - built-in

# Optional: Enhanced logging and monitoring
# colorlog>=6.7.0
# psutil>=5.9.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
